refactor(tests): split editor tests and document shared doubles
- Split EditMode tests by feature and moved test doubles to a shared asmdef outside the Editor folder to avoid the AddComponent restriction.
- Updated test structure documentation in README/CLAUDE to match the implementation.
